backend/pyproject.toml
[project]
name = "financial-data-backend"
version = "1.0.0"
description = "Financial Data Dashboard Backend API"
authors = [
  { name = "AHMED Moorsy", email = "ahmed.moorsy798@gmail.com" }
]
license = "MIT"
requires-python = ">=3.11,<3.13"

dependencies = [
  "fastapi[standard]>=0.115.0",
  "uvicorn>=0.31.0",
  "sqlalchemy>=2.0.0",
  "psycopg2-binary>=2.9.0",
  "pydantic>=2.0.0",
  "pydantic-settings>=2.0.0",
  "copilotkit==0.1.67",
  "langgraph==0.6.10",
  "langchain-core>=0.3.79",
  "langchain-community>=0.3.31",
  "langchain==0.3.27",
  "langchain-openai>=0.3.35",
  "openai>=2.3.0",
  "python-dotenv>=1.0.1",
  "ag-ui-protocol>=0.1.7",
]

[dependency-groups]
dev = [
  "pytest>=8.0.0",
  "pytest-asyncio>=0.23.0",
  "pytest-cov>=4.1.0",
  "httpx>=0.27.0",
  "faker>=24.0.0",
  "pytest-mock>=3.12.0",
]

[project.scripts]
start = "src.main:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
